Retire product_contents_compiler; fold status.txt + deck_map into contents_validator
MTGJSON's pipeline (mtgjson5/pipeline/stages/sealed.py) now compiles the
sealed contents directly from this repo's raw YAML (it explicitly
'Replicates' new_products_compiler / product_contents_compiler /
card_to_product_compiler and reads data/products + data/contents from the
repo tarball). So outputs/contents.json is unconsumed downstream.

product_contents_compiler.py was still doing two useful things, though:
  - writing status.txt (the coverage report), and
  - writing outputs/deck_map.json, which import_new_decks.py reads.

Move both into contents_validator.py behind a --status flag (build_uuid_map
and deck_links are copied verbatim). Default 'contents_validator.py' stays
the fast structural PR check with no new deps (ijson/requests are lazy-
imported only in --status mode). Repoint the rebuild/daily/build_caches
workflows from product_contents_compiler.py to 'contents_validator.py
--status', delete the compiler, and drop the now-unconsumed contents.json.

card_to_product_compiler.py (outputs/card_map.json) and the token/gatherer
mappings are untouched -- those outputs are still consumed.
