+++ .github/workflows/build_caches.yml
@@ -25,8 +25,8 @@ jobs:
           python -m pip install --upgrade pip
           pip install -r requirements.txt
 
-      - name: execute contents compiler
-        run: python scripts/product_contents_compiler.py
+      - name: rebuild status + deck map
+        run: python scripts/contents_validator.py --status
 
       - name: execute manual products compiler
         run: python scripts/new_products_compiler.py
