+++ .github/workflows/rebuild_outputs.yml
@@ -35,8 +35,8 @@ jobs:
       - name: execute import of new deck
         run: python scripts/import_new_decks.py
 
-      - name: execute contents compiler
-        run: python scripts/product_contents_compiler.py
+      - name: rebuild status + deck map
+        run: python scripts/contents_validator.py --status
 
       - name: execute manual products compiler
         run: python scripts/new_products_compiler.py
