--- outputs/contents.json
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:f0032fd154f081be382de423489bf483188382740100f56d5b93e4ed05461a9b
-size 1892886
