+++ scripts/contents_validator.py
@@ -1,6 +1,148 @@
+"""Validate sealed-product YAML, and (with --status) rebuild status.txt + deck_map.json.
+
+This absorbs the status-report and deck-map role of product_contents_compiler.py so
+that compiler can be retired:
+
+  * status.txt      -- the per-product coverage report contributors rely on
+  * deck_map.json   -- deck-to-product UUID mapping
+
+The old compiler also emitted outputs/contents.json, but MTGJSON's own pipeline
+(mtgjson5/pipeline/stages/sealed.py) now compiles the sealed contents directly from
+this repo's raw YAML, so that output is unconsumed and is no longer produced here.
+
+    python scripts/contents_validator.py            # fast structural validation (CI PR gate)
+    python scripts/contents_validator.py --status   # also rebuild status.txt + deck_map.json
+"""
+import argparse
+import json
 import yaml
 import product_classes as pc
 from pathlib import Path
+
+def build_uuid_map(mtgjson_path):
+    import ijson
+    import requests
+    print("🤖 loading mtgjson...")
+    try:
+        if mtgjson_path:
+            print("⚙️  using local AllPrintings.json")
+            f = open(mtgjson_path, 'rb')
+            parser = ijson.parse(f)
+        else:
+            print("⚙️  downloading AllPrintings.json")
+            url = "https://mtgjson.com/api/v5/AllPrintings.json"
+            r = requests.get(url, stream=True)
+            parser = ijson.parse(r.content)
+    except:
+        print("Could not load AllPrintings")
+        return
+
+    print("🤖 filtering sealed products...")
+
+    uuids = {}
+    current_set = ""
+    status = ""
+    name = ""
+    number = ""
+    uuid = ""
+    holding = ""
+    for prefix, event, value in parser:
+        if prefix == "data" and event == "map_key":
+            current_set = value
+            ccode = current_set.lower()
+            uuids[ccode] = {
+                "booster": set(),
+                "decks": set(),
+                "sealedProduct": {},
+                "cards": {},
+                "tokens": {},
+            }
+            status = ""
+        elif prefix == f"data.{current_set}" and event == "map_key":
+            status = value
+        elif (
+            status == "booster"
+            and prefix == f"data.{current_set}.booster"
+            and event == "map_key"
+        ):
+            uuids[ccode]["booster"].add(value)
+        elif status == "decks" and prefix == f"data.{current_set}.decks.item.name":
+            uuids[ccode]["decks"].add(value)
+        elif status == "sealedProduct":
+            if (
+                prefix == f"data.{current_set}.sealedProduct.item"
+                and event == "start_map"
+            ):
+                name = ""
+                uuid = ""
+            elif prefix == f"data.{current_set}.sealedProduct.item.name":
+                name = value
+            elif prefix == f"data.{current_set}.sealedProduct.item.uuid":
+                uuid = value
+            elif (
+                prefix == f"data.{current_set}.sealedProduct.item"
+                and event == "end_map"
+            ):
+                uuids[ccode]["sealedProduct"][name] = uuid
+        elif status == "cards":
+            # Only preserve the "main" face of the card
+            if prefix == f"data.{current_set}.cards.item.side":
+                if value != "a":
+                    holding = "skip"
+            if prefix == f"data.{current_set}.cards.item" and event == "start_map":
+                number = ""
+                name = ""
+                uuid = ""
+            elif prefix == f"data.{current_set}.cards.item.number":
+                number = value
+            elif prefix == f"data.{current_set}.cards.item.name":
+                name = value
+            elif prefix == f"data.{current_set}.cards.item.uuid":
+                uuid = value
+            elif prefix == f"data.{current_set}.cards.item" and event == "end_map":
+                if holding != "skip":
+                    uuids[ccode]["cards"][number] = (uuid, name)
+                holding = ""
+        elif status == "tokens":
+            # Tokens live in a separate array from cards but are referenced the
+            # same way (by collector number), e.g. SLD 918 "Food". Index them so
+            # card lookups can fall back here. Only keep the "main" face.
+            if prefix == f"data.{current_set}.tokens.item.side":
+                if value != "a":
+                    holding = "skip"
+            if prefix == f"data.{current_set}.tokens.item" and event == "start_map":
+                number = ""
+                name = ""
+                uuid = ""
+            elif prefix == f"data.{current_set}.tokens.item.number":
+                number = value
+            elif prefix == f"data.{current_set}.tokens.item.name":
+                name = value
+            elif prefix == f"data.{current_set}.tokens.item.uuid":
+                uuid = value
+            elif prefix == f"data.{current_set}.tokens.item" and event == "end_map":
+                if holding != "skip":
+                    uuids[ccode]["tokens"][number] = (uuid, name)
+                holding = ""
+
+    if mtgjson_path:
+        f.close()
+
+    return uuids
+
+def deck_links(all_products):
+    deck_mapper = {}
+    for set_contents in all_products.values():
+        for product_contents in set_contents.values():
+            if not product_contents.uuid:
+                continue
+            for deck in product_contents.deck:
+                if deck.set not in deck_mapper:
+                    deck_mapper[deck.set] = {}
+                if deck.name not in deck_mapper[deck.set]:
+                    deck_mapper[deck.set][deck.name] = []
+                deck_mapper[deck.set][deck.name].append(product_contents.uuid)
+    return deck_mapper
 
 valid_categories = [
     "BOOSTER_PACK", "BOOSTER_BOX", "BOOSTER_CASE", "DECK", "MULTI_DECK",
@@ -19,7 +161,7 @@ valid_subtypes = [
     "SEALED_SET", "PRERELEASE", "OTHER", "CHALLENGER", "SIX", "CONVENTION", "MTGO_REDEMPTION",
 ]
 
-if __name__ == "__main__":
+def validate_structure():
     contentFolder = Path("data/contents/")
     failed = False
     for set_file in contentFolder.glob("*.yaml"):
@@ -49,7 +191,7 @@ if __name__ == "__main__":
     for set_file in all_files:
         with open(set_file, "rb") as f:
             contents = yaml.safe_load(f)
-        
+
         for name, p in contents["products"].items():
             if "category" not in p.keys():
                 print(f"Product {name} in set {set_file.stem} missing category")
@@ -74,3 +216,60 @@ if __name__ == "__main__":
     if failed:
         raise ImportError()
     print("All products validated")
+
+def rebuild_status_and_deck_map(uuid_map):
+    """Rebuild status.txt (coverage) and outputs/deck_map.json.
+
+    Same product compilation the old product_contents_compiler.py ran -- product
+    construction and get_uuids() append the "missing contents" / "not found"
+    lines to status.txt -- but without emitting the unconsumed contents.json."""
+    products_contents = {}
+    status_file = Path("status.txt")
+    with open(status_file, "w") as f:
+        f.write("Starting output\n")
+
+    for set_file in sorted(Path("data/contents/").glob("*.yaml")):
+        with open(set_file, "rb") as f:
+            contents = yaml.safe_load(f)
+
+        products_contents[contents["code"]] = {}
+        for name, p in contents["products"].items():
+            if not p:
+                with open(status_file, "a") as f:
+                    f.write(f"Product {contents['code']} - {name} missing contents\n")
+                continue
+            if set(p.keys()) == {"copy"}:
+                p = contents["products"][p["copy"]]
+            compiled_product = pc.product(p, contents["code"], name)
+            compiled_product.get_uuids(uuid_map)
+            products_contents[contents["code"]][name] = compiled_product
+        if not products_contents[contents["code"]]:
+            products_contents.pop(contents["code"])
+
+    deck_map = deck_links(products_contents)
+    with open("outputs/deck_map.json", "w") as outfile:
+        json.dump(deck_map, outfile)
+
+def parse_args() -> argparse.Namespace:
+    parser = argparse.ArgumentParser("contents_validator")
+    parser.add_argument(
+        "--status", action="store_true",
+        help="rebuild status.txt + deck_map.json (loads MTGJSON AllPrintings)",
+    )
+    parser.add_argument(
+        "--mtgjson", "-m", type=str, required=False,
+        help="path to a local AllPrintings.json (downloaded if omitted)",
+    )
+    return parser.parse_args()
+
+
+if __name__ == "__main__":
+    args = parse_args()
+    if args.status:
+        uuid_map = build_uuid_map(args.mtgjson)
+        if not uuid_map:
+            raise SystemExit("Aborting: AllPrintings could not be loaded")
+        rebuild_status_and_deck_map(uuid_map)
+        print("🤖 status.txt and deck_map.json written")
+    else:
+        validate_structure()
