--- scripts/product_contents_compiler.py
@@ -1,189 +1,0 @@
-import argparse
-import yaml
-import product_classes as pc
-from pathlib import Path
-import json
-import ijson
-import requests
-
-
-def set_to_json(set_content):
-    decoded = {k: v.toJson() for k, v in set_content.items()}
-    return {k: v for k, v in decoded.items() if v}
-
-
-def build_uuid_map(mtgjson_path):
-    print("🤖 loading mtgjson...")
-    try:
-        if mtgjson_path:
-            print("⚙️  using local AllPrintings.json")
-            f = open(mtgjson_path, 'rb')
-            parser = ijson.parse(f)
-        else:
-            print("⚙️  downloading AllPrintings.json")
-            url = "https://mtgjson.com/api/v5/AllPrintings.json"
-            r = requests.get(url, stream=True)
-            parser = ijson.parse(r.content)
-    except:
-        print("Could not load AllPrintings")
-        return
-
-    print("🤖 filtering sealed products...")
-
-    uuids = {}
-    current_set = ""
-    status = ""
-    name = ""
-    number = ""
-    uuid = ""
-    holding = ""
-    for prefix, event, value in parser:
-        if prefix == "data" and event == "map_key":
-            current_set = value
-            ccode = current_set.lower()
-            uuids[ccode] = {
-                "booster": set(),
-                "decks": set(),
-                "sealedProduct": {},
-                "cards": {},
-                "tokens": {},
-            }
-            status = ""
-        elif prefix == f"data.{current_set}" and event == "map_key":
-            status = value
-        elif (
-            status == "booster"
-            and prefix == f"data.{current_set}.booster"
-            and event == "map_key"
-        ):
-            uuids[ccode]["booster"].add(value)
-        elif status == "decks" and prefix == f"data.{current_set}.decks.item.name":
-            uuids[ccode]["decks"].add(value)
-        elif status == "sealedProduct":
-            if (
-                prefix == f"data.{current_set}.sealedProduct.item"
-                and event == "start_map"
-            ):
-                name = ""
-                uuid = ""
-            elif prefix == f"data.{current_set}.sealedProduct.item.name":
-                name = value
-            elif prefix == f"data.{current_set}.sealedProduct.item.uuid":
-                uuid = value
-            elif (
-                prefix == f"data.{current_set}.sealedProduct.item"
-                and event == "end_map"
-            ):
-                uuids[ccode]["sealedProduct"][name] = uuid
-        elif status == "cards":
-            # Only preserve the "main" face of the card
-            if prefix == f"data.{current_set}.cards.item.side":
-                if value != "a":
-                    holding = "skip"
-            if prefix == f"data.{current_set}.cards.item" and event == "start_map":
-                number = ""
-                name = ""
-                uuid = ""
-            elif prefix == f"data.{current_set}.cards.item.number":
-                number = value
-            elif prefix == f"data.{current_set}.cards.item.name":
-                name = value
-            elif prefix == f"data.{current_set}.cards.item.uuid":
-                uuid = value
-            elif prefix == f"data.{current_set}.cards.item" and event == "end_map":
-                if holding != "skip":
-                    uuids[ccode]["cards"][number] = (uuid, name)
-                holding = ""
-        elif status == "tokens":
-            # Tokens live in a separate array from cards but are referenced the
-            # same way (by collector number), e.g. SLD 918 "Food". Index them so
-            # card lookups can fall back here. Only keep the "main" face.
-            if prefix == f"data.{current_set}.tokens.item.side":
-                if value != "a":
-                    holding = "skip"
-            if prefix == f"data.{current_set}.tokens.item" and event == "start_map":
-                number = ""
-                name = ""
-                uuid = ""
-            elif prefix == f"data.{current_set}.tokens.item.number":
-                number = value
-            elif prefix == f"data.{current_set}.tokens.item.name":
-                name = value
-            elif prefix == f"data.{current_set}.tokens.item.uuid":
-                uuid = value
-            elif prefix == f"data.{current_set}.tokens.item" and event == "end_map":
-                if holding != "skip":
-                    uuids[ccode]["tokens"][number] = (uuid, name)
-                holding = ""
-
-    if mtgjson_path:
-        f.close()
-
-    return uuids
-
-
-def deck_links(all_products):
-    deck_mapper = {}
-    for set_contents in all_products.values():
-        for product_contents in set_contents.values():
-            if not product_contents.uuid:
-                continue
-            for deck in product_contents.deck:
-                if deck.set not in deck_mapper:
-                    deck_mapper[deck.set] = {}
-                if deck.name not in deck_mapper[deck.set]:
-                    deck_mapper[deck.set][deck.name] = []
-                deck_mapper[deck.set][deck.name].append(product_contents.uuid)
-    return deck_mapper
-
-
-def main(args: argparse.Namespace):
-    uuid_map = build_uuid_map(args.mtgjson)
-    if not uuid_map:
-        raise SystemExit("Aborting: AllPrintings could not be loaded")
-    products_contents = {}
-    status_file = Path("status.txt")
-    with open(status_file, "w") as f:
-        f.write("Starting output\n")
-
-    print("🤖 loading existing products files...")
-
-    for set_file in sorted(Path("data/contents/").glob("*.yaml")):
-        with open(set_file, "rb") as f:
-            contents = yaml.safe_load(f)
-
-        products_contents[contents["code"]] = {}
-        for name, p in contents["products"].items():
-            if not p:
-                with open(status_file, "a") as f:
-                    f.write(f"Product {contents['code']} - {name} missing contents\n")
-                continue
-            if set(p.keys()) == {"copy"}:
-                p = contents["products"][p["copy"]]
-            compiled_product = pc.product(p, contents["code"], name)
-            compiled_product.get_uuids(uuid_map)
-            products_contents[contents["code"]][name] = compiled_product
-        if not products_contents[contents["code"]]:
-            products_contents.pop(contents["code"])
-
-    print("🤖 saving results...")
-
-    with open("outputs/contents.json", "w") as outfile:
-        json.dump({k: set_to_json(v) for k, v in products_contents.items()}, outfile)
-
-    deck_map = deck_links(products_contents)
-
-    with open("outputs/deck_map.json", "w") as outfile:
-        json.dump(deck_map, outfile)
-
-
-def parse_args() -> argparse.Namespace:
-    parser = argparse.ArgumentParser("product_contents_compiler")
-
-    parser.add_argument("--mtgjson", "-m", type=str, required=False)
-
-    return parser.parse_args()
-
-
-if __name__ == "__main__":
-    main(parse_args())
